requirements.txt
pyrogram==2.0.106
motor==3.3.2
pymongo==4.6.1
TgCrypto==1.2.5
dnspython==2.4.2
psutil==5.9.6
requests==2.31.0
python-decouple==3.8
asyncio
colorama==0.4.6
rich==13.7.0